📝 Update monoblock examples.
Update monoblock examples with the following changes:
* Update description comments to more clearly describe the models. In
particular clarify that the central cylinder in the simplified model
represents the interlayer and not the pipe as previously described.
* As above, update variable names referencing the pipe in the simple
model to instead refer to the interlayer.
* Update file handling names and add output directory variable.
* Add lines to the simple model to explicitly set the mesh to first
order.
* Minor structure fixes for consistency between files.
* Add monoBThick variable to more consistently describe the
monoblock geometry with respect to the radial build. Change monoBWidth
to be a calculated value. Rename the previously existing monoBThick
(referring to the z-dimension) to monoBDepth to avoid a namespace clash.
* Add lines to calculated values for number of divisions in the mesh
ensuring that these values are rounded to an even integer. Add comments
explaining the need for these values to be even integers.
* Reorder material properties in both examples to sort primarily by
parameter and secondarily by material, improving readability.
* Correct block assignment of cucrzr and copper material properties to
the pipe and interlayer respectively, which were previously assigned
incorrectly.